add util and people module

+++ common/src/country.rs
@@ -8,22 +8,22 @@ use serde::{Serialize, Deserialize};
 
 #[derive(Serialize, Deserialize)]
 pub struct Country {
-    id: u8,
+    id: u32,
     name: String,
     alpha_2: String,
     alpha_3: String,
-    currency_id: u8
+    currency_id: u32
 }
 
 
 impl Country {
 
     pub fn new(
-        id: &u8,
+        id: &u32,
         name: &str,
         alpha_2: &str,
         alpha_3: &str,
-        currency_id: &u8
+        currency_id: &u32
     ) -> Self {
         return Self {
             id: id.clone(),
@@ -34,7 +34,7 @@ impl Country {
         };
     }
 
-    pub fn id(&self) -> u8 {
+    pub fn id(&self) -> u32 {
         return self.id.clone();
     }
 
@@ -50,7 +50,7 @@ impl Country {
         return self.alpha_3.clone();
     }
 
-    pub fn currency_id(&self) -> u8 {
+    pub fn currency_id(&self) -> u32 {
         return self.currency_id.clone();
     }
 }
